feat(mcp): extract MCP server wiring into server-lib
Move SDK server construction and request handler registration into a
focused lib module with a thin re-export wrapper and targeted tests.
